split add styles input
